--- components/CategoryManager/Notification.js
@@ -1,39 +1,0 @@
-import React from 'react';
-import { Snackbar } from '@mui/material';
-import styled from 'styled-components';
-import { Alert } from '@mui/lab';
-
-const StySnackbar = styled(Snackbar)`
-  root: {
-    top: 40px;
-  }
-`;
-
-const Notification = (props) => {
-  const { notify, setNotify } = props;
-
-  const handleClose = (reason) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-    setNotify({
-      ...notify,
-      isOpen: false,
-    });
-  };
-
-  return (
-    <StySnackbar
-      open={notify.isOpen}
-      autoHideDuration={3000}
-      anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-      onClose={handleClose}
-    >
-      <Alert severity={notify.type} onClose={handleClose}>
-        {notify.message}
-      </Alert>
-    </StySnackbar>
-  );
-};
-
-export default Notification;
