--- components/CategoryManager/controls/Controls.js
@@ -1,11 +1,0 @@
-import Input from './Input';
-import Button from './Button';
-import ActionButton from './ActionButton';
-
-const Controls = {
-  Input,
-  Button,
-  ActionButton,
-};
-
-export default Controls;
