+++ services/Post.js
@@ -1,5 +1,8 @@
 import http from '../utils/axios';
 
+export const getPostByTitleContainingString = (string) =>
+  http(`/v1/post/search/${string}`, { method: 'GET' });
+
 export const postPost = ({ categoryTitle, title, context, headImgUrl }) =>
   http(`/v1/post`, {
     method: 'POST',
@@ -20,9 +23,6 @@ export const editPost = ({ context, headImgUrl, title }, postId) =>
       title,
     },
   });
-
-export const getPostByTitleContainingString = (string) =>
-  http(`/v1/post/string/${string}`, { method: 'GET' });
 
 export const getFollowPostByUsername = (username) =>
   http(`/v1/post/follows/find_all_by_username/${username}`, { method: 'GET' });
